+++ CHANGES.md
@@ -8,6 +8,7 @@
 * Remove `VariableMapping.merge_instructions_for`.
 * Add optional options `:input_id` and `:output_id` to `VariableMapping.merge_instructions_for`. You don't need this, though.
 * Remove `Insert.find`.
+* Move `Pipeline.prepend` to the `trailblazer-activity-linear-dsl` gem.
 
 # 0.13.0